rockusb: Deduplicate example code
When the nusb backend was added, most of the rockusb example code was
simply copied and made async. Which wasn't a bit problem, but it causes
both examples to deviate when new features are added. To prevent that
refactor the examples to primarily use common helper code for most
commands, apart from the transport specific parts.

This does drop the nbd command as there is no nice async crate for it
and really it was mostly a cool gimmick